refactor: removed `basic` example from tui-editor
