+++ projects/demo-integrations/cypress/support/editor/html.ts
@@ -1,0 +1,128 @@
+export const HTML_EDITOR_BASIC_EXAMPLE = `
+<p>
+    <strong>WYSIWYG</strong>
+    (
+    <em>What you see is what you get</em>
+    ) —
+    <u>Rich Text Editor</u>
+    for using with Angular
+    <sup>forms.</sup>
+</p>
+<h1>Heading</h1>
+<p>
+    Lorem ipsum dolor sit amet consectetur
+    <a
+        target="_blank"
+        rel="noopener noreferrer nofollow"
+        href="http://taiga-ui.dev"
+    >
+        adipiscing elit
+    </a>
+    , sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim
+</p>
+<blockquote>
+    <p>
+        ad minim veniam, quis nostrud exercitation
+        <span style="color: rgb(196, 11, 8); background-color: rgb(221, 228, 237)">ullamco</span>
+        , laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit
+        esse cillum dolore eu fugiat nulla pariatur.
+    </p>
+</blockquote>
+<p style="text-align: right">
+    Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
+</p>
+<pre><code>class EditorExample {}</code></pre>
+<table>
+    <tbody>
+        <tr>
+            <th
+                colspan="1"
+                rowspan="1"
+            >
+                <p></p>
+            </th>
+            <th
+                colspan="1"
+                rowspan="1"
+            >
+                <p>Free</p>
+            </th>
+            <th
+                colspan="1"
+                rowspan="1"
+            >
+                <p>Pro</p>
+            </th>
+        </tr>
+        <tr>
+            <td
+                colspan="1"
+                rowspan="1"
+            >
+                <p>24/7 support</p>
+            </td>
+            <td
+                style="background: rgb(255, 221, 41)"
+                colspan="1"
+                rowspan="1"
+            >
+                <p>+</p>
+            </td>
+            <td
+                style="background: rgb(57, 181, 77)"
+                colspan="1"
+                rowspan="1"
+            >
+                <p>+</p>
+            </td>
+        </tr>
+    </tbody>
+</table>
+<p style="text-align: center"><code>Code in text</code></p>
+`;
+
+export const HTML_EDITOR_EXAMPLE_UL = `
+    <ul>
+        <li>1</li>
+        <li>2</li>
+    </ul>
+
+    <ol>
+        <li>A</li>
+        <li>B</li>
+    </ol>
+`;
+
+export const HTML_EDITOR_EXAMPLE_NESTED_UL = `
+<ul>
+  <li>
+    1
+    <ul>
+      <li>
+        2
+        <ul>
+          <li>3</li>
+          <li>4</li>
+        </ul>
+      </li>
+    </ul>
+  </li>
+  <li>5</li>
+</ul>
+
+<ol>
+  <li>
+    1
+    <ol>
+      <li>
+        2
+        <ol>
+          <li>3</li>
+          <li>4</li>
+        </ol>
+      </li>
+    </ol>
+  </li>
+  <li>5</li>
+</ol>
+`;
