--- projects/demo-integrations/cypress/tests/addon-editor/utils.ts
@@ -1,1 +1,0 @@
-export const WAIT_BEFORE_SCREENSHOT = 1000; // editor has a huge number of icons (that is why we wait so long)
