+++ projects/demo/src/modules/components/editor-new/editor-new.template.html
@@ -32,19 +32,10 @@
         </tui-notification>
 
         <tui-doc-example
-            id="basic"
-            i18n-heading
-            heading="Basic"
-            [content]="example1"
-        >
-            <tui-editor-example-1></tui-editor-example-1>
-        </tui-doc-example>
-
-        <tui-doc-example
             id="custom-tool"
             i18n-heading
-            heading="With custom tool"
-            [content]="example2"
+            heading="Custom tool"
+            [content]="example1"
         >
             You can create your own tool:
             <ul
@@ -74,34 +65,34 @@
                     .
                 </li>
             </ul>
-            <tui-editor-example-2></tui-editor-example-2>
+            <tui-editor-example-1></tui-editor-example-1>
         </tui-doc-example>
 
         <tui-doc-example
             id="resizable-image"
             i18n-heading
-            heading="With resizable image"
+            heading="Resizable image"
+            [content]="example2"
+        >
+            <tui-editor-example-2></tui-editor-example-2>
+        </tui-doc-example>
+
+        <tui-doc-example
+            id="preview-image"
+            i18n-heading
+            heading="Preview for images"
             [content]="example3"
         >
             <tui-editor-example-3></tui-editor-example-3>
         </tui-doc-example>
 
         <tui-doc-example
-            id="preview-image"
-            i18n-heading
-            heading="Show preview for images"
-            [content]="example4"
-        >
-            <tui-editor-example-4></tui-editor-example-4>
-        </tui-doc-example>
-
-        <tui-doc-example
             id="custom-processing"
             i18n-heading
             heading="Custom processing content"
-            [content]="example5"
+            [content]="example4"
         >
-            <tui-editor-example-5></tui-editor-example-5>
+            <tui-editor-example-4></tui-editor-example-4>
         </tui-doc-example>
     </ng-template>
 
@@ -109,6 +100,7 @@
         <tui-doc-demo>
             <tui-editor
                 new
+                [style.max-height.px]="maxHeight"
                 [formControl]="control"
                 [focusable]="focusable"
                 [exampleText]="exampleText"
@@ -120,6 +112,12 @@
             >
                 Start typing
             </tui-editor>
+
+            <h4>HTML:</h4>
+            <tui-editor-socket [content]="control.value"></tui-editor-socket>
+
+            <h4>Text:</h4>
+            <p>{{ control.value }}</p>
         </tui-doc-demo>
         <tui-doc-documentation>
             <ng-template
@@ -148,6 +146,15 @@
                 documentationPropertyMode="input"
                 documentationPropertyType="string"
                 [(documentationPropertyValue)]="exampleText"
+            >
+                Example text shown on empty focused input
+            </ng-template>
+            <ng-template
+                i18n
+                documentationPropertyName="max-height"
+                documentationPropertyMode="input"
+                documentationPropertyType="string"
+                [(documentationPropertyValue)]="maxHeight"
             >
                 Example text shown on empty focused input
             </ng-template>
