+++ projects/demo/src/modules/components/editor-new/examples/1/index.less
@@ -1,3 +1,6 @@
-.editor {
-    max-height: 18.75rem;
+.hint {
+    color: var(--tui-base-05);
+    height: 100%;
+    display: flex;
+    align-items: center;
 }
