+++ projects/demo/src/modules/components/editor-new/examples/1/smiles-tool/smiles-tool.styles.less
