+++ projects/demo/src/modules/components/editor-new/examples/2/index.html
@@ -3,15 +3,7 @@
     class="editor"
     [formControl]="control"
     [tools]="builtInTools"
->
-    Smiles are custom tool. Try it.
+></tui-editor>
 
-    <ng-container ngProjectAs="tools">
-        <smiles-tool tuiToolbarTool></smiles-tool>
-
-        <span class="hint">
-            <tui-svg src="tuiIconArrowLeft"></tui-svg>
-            click it
-        </span>
-    </ng-container>
-</tui-editor>
+<h4>HTML:</h4>
+<tui-editor-socket [content]="control.value"></tui-editor-socket>
