+++ projects/demo/src/modules/components/editor-new/examples/2/index.less
@@ -1,6 +1,3 @@
-.hint {
-    color: var(--tui-base-05);
-    height: 100%;
-    display: flex;
-    align-items: center;
+.editor {
+    min-height: 30rem;
 }
