+++ projects/demo/src/modules/components/editor-new/examples/3/image-preview/image-preview.template.html
