+++ projects/demo/src/modules/components/editor-new/examples/3/index.html
@@ -6,4 +6,9 @@
 ></tui-editor>
 
 <h4>HTML:</h4>
-<tui-editor-socket [content]="control.value"></tui-editor-socket>
+<tui-editor-socket
+    [content]="control.value"
+    (imagePreview)="preview.showImage($event)"
+>
+    <image-preview-example #preview></image-preview-example>
+</tui-editor-socket>
