+++ projects/demo/src/modules/components/editor-new/examples/4/index.html
@@ -2,13 +2,12 @@
     new
     class="editor"
     [formControl]="control"
-    [tools]="builtInTools"
-></tui-editor>
+>
+    Placeholder
+</tui-editor>
 
 <h4>HTML:</h4>
-<tui-editor-socket
-    [content]="control.value"
-    (imagePreview)="preview.showImage($event)"
->
-    <image-preview-example #preview></image-preview-example>
-</tui-editor-socket>
+<tui-editor-socket [content]="control.value"></tui-editor-socket>
+
+<h4>Text:</h4>
+<p>{{ control.value }}</p>
