+++ projects/demo/src/modules/components/editor-new/examples/4/index.less
@@ -1,3 +1,3 @@
 .editor {
-    min-height: 30rem;
+    min-height: 20rem;
 }
