--- projects/demo/src/modules/components/editor-new/examples/5/index.html
@@ -1,13 +1,0 @@
-<tui-editor
-    new
-    class="editor"
-    [formControl]="control"
->
-    Placeholder
-</tui-editor>
-
-<h4>HTML:</h4>
-<tui-editor-socket [content]="control.value"></tui-editor-socket>
-
-<h4>Text:</h4>
-<p>{{ control.value }}</p>
