--- projects/demo/src/modules/components/editor-new/examples/5/index.less
@@ -1,3 +1,0 @@
-.editor {
-    min-height: 20rem;
-}
